refactor: topic thumbs lib to topics.thumbs
